+++ rex/renewable_resource.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 import warnings
 
-from rex.resource import Resource, MultiFileResource
+from rex.resource import Resource
 from rex.sam_resource import SAMResource
 from rex.utilities.exceptions import (ResourceValueError, ExtrapolationWarning,
                                       ResourceWarning,
@@ -1115,185 +1115,5 @@ class WaveResource(Resource):
                   "str_decode": str_decode, "group": group}
         with cls(h5_file, **kwargs) as res:
             SAM_res = res._preload_SAM(sites, means=means)
-
-        return SAM_res
-
-
-class MultiFileNSRDB(MultiFileResource, NSRDB):
-    """
-    Class to handle 2018 and beyond NSRDB data that is at 2km and
-    sub 30 min resolution
-
-    See Also
-    --------
-    resource.MultiFileResource : Parent class
-    resource.NSRDB : Parent class
-    """
-    @classmethod
-    def preload_SAM(cls, h5_source, sites, unscale=True, str_decode=True,
-                    tech='pvwattsv7', clearsky=False, bifacial=False,
-                    downscale=None, means=False, check_file=False):
-        """
-        Pre-load project_points for SAM
-
-        Parameters
-        ----------
-        h5_source : str | list
-            Path to directory containing multi-file resource file sets.
-            Available formats:
-                /h5_dir/
-                /h5_dir/prefix*suffix
-            Or list of source .h5 files
-        sites : list
-            List of sites to be provided to SAM
-        unscale : bool
-            Boolean flag to automatically unscale variables on extraction
-        str_decode : bool
-            Boolean flag to decode the bytestring meta data into normal
-            strings. Setting this to False will speed up the meta data read.
-        tech : str, optional
-            SAM technology string, by default 'pvwattsv7'
-        clearsky : bool
-            Boolean flag to pull clearsky instead of real irradiance
-        bifacial : bool
-            Boolean flag to pull surface albedo for bifacial modeling.
-        downscale : NoneType | str
-            Option for NSRDB resource downscaling to higher temporal
-            resolution. Expects a string in the Pandas frequency format,
-            e.g. '5min'.
-        means : bool
-            Boolean flag to compute mean resource when res_array is set
-        check_files : bool
-            Check to ensure files have the same coordinates and time_index
-
-        Returns
-        -------
-        SAM_res : SAMResource
-            Instance of SAMResource pre-loaded with Solar resource for sites
-            in project_points
-        """
-        with cls(h5_source, unscale=unscale, str_decode=str_decode,
-                 check_file=check_file) as res:
-            # pylint: disable=assignment-from-no-return
-            SAM_res = res._preload_SAM(sites, tech=tech, clearsky=clearsky,
-                                       bifacial=bifacial, downscale=downscale,
-                                       means=means)
-
-        return SAM_res
-
-
-class MultiFileWTK(MultiFileResource, WindResource):
-    """
-    Class to handle 5min WIND Toolkit data
-
-    See Also
-    --------
-    resource.MultiFileResource : Parent class
-    resource.WindResource : Parent class
-
-    Examples
-    --------
-    MultiFileWTK automatically searches for files of the form *m.h5
-
-    >>> file = '$TESTDATADIR/wtk'
-    >>> with MultiFileWTK(file) as res:
-    >>>     print(list(res._h5_files)
-    >>>     print(res.datasets)
-    ['$TESTDATADIR/wtk_2010_200m.h5',
-     '$TESTDATADIR/wtk_2010_100m.h5']
-    ['coordinates', 'meta', 'pressure_100m', 'pressure_200m',
-     'temperature_100m', 'temperature_200m', 'time_index',
-     'winddirection_100m', 'winddirection_200m', 'windspeed_100m',
-     'windspeed_200m']
-
-    MultiFileWTK, like WindResource can interpolate / extrapolate hub-heights
-
-    >>> with MultiFileWTK(file) as res:
-    >>>     wspd = res['windspeed_150m']
-    >>>
-    >>> wspd
-    [[16.19     16.25     16.305    ... 16.375    16.39     16.39    ]
-     [16.15     16.205    16.255001 ... 16.35     16.365    16.39    ]
-     [16.154999 16.195    16.23     ... 16.335    16.32     16.34    ]
-     ...
-     [10.965    11.675    12.08     ... 15.18     14.805    14.42    ]
-     [11.66     11.91     12.535    ... 13.31     12.23     12.335   ]
-     [12.785    13.295    14.014999 ... 12.205    11.360001 11.64    ]]
-    """
-    SUFFIX = 'm.h5'
-
-    def __init__(self, h5_source, unscale=True, str_decode=True,
-                 check_files=False):
-        """
-        Parameters
-        ----------
-        h5_source : str | list
-            Path to directory containing multi-file resource file sets.
-            Available formats:
-                /h5_dir/
-                /h5_dir/prefix*suffix
-            Or list of source .h5 files
-        unscale : bool
-            Boolean flag to automatically unscale variables on extraction
-        str_decode : bool
-            Boolean flag to decode the bytestring meta data into normal
-            strings. Setting this to False will speed up the meta data read.
-        check_files : bool
-            Check to ensure files have the same coordinates and time_index
-        """
-        super().__init__(h5_source, unscale=unscale, str_decode=str_decode,
-                         check_files=check_files)
-        self._heights = None
-
-    @classmethod
-    def preload_SAM(cls, h5_source, sites, hub_heights, unscale=True,
-                    str_decode=True, require_wind_dir=False,
-                    precip_rate=False, icing=False, means=False,
-                    check_files=False):
-        """
-        Placeholder for classmethod that will pre-load project_points for SAM
-
-        Parameters
-        ----------
-        h5_source : str | list
-            Path to directory containing multi-file resource file sets.
-            Available formats:
-                /h5_dir/
-                /h5_dir/prefix*suffix
-            Or list of source .h5 files
-        sites : list
-            List of sites to be provided to SAM
-        hub_heights : int | float | list
-            Hub heights to extract for SAM
-        unscale : bool
-            Boolean flag to automatically unscale variables on extraction
-        str_decode : bool
-            Boolean flag to decode the bytestring meta data into normal
-            strings. Setting this to False will speed up the meta data read.
-        require_wind_dir : bool
-            Boolean flag as to whether wind direction will be loaded.
-        precip_rate : bool
-            Boolean flag as to whether precipitationrate_0m will be preloaded
-        icing : bool
-            Boolean flag as to whether icing is analyzed.
-            This will preload relative humidity.
-        means : bool
-            Boolean flag to compute mean resource when res_array is set
-        check_files : bool
-            Check to ensure files have the same coordinates and time_index
-
-        Returns
-        -------
-        SAM_res : SAMResource
-            Instance of SAMResource pre-loaded with Solar resource for sites
-            in project_points
-        """
-        with cls(h5_source, unscale=unscale, str_decode=str_decode,
-                 check_files=check_files) as res:
-            # pylint: disable=assignment-from-no-return
-            SAM_res = res._preload_SAM(sites, hub_heights,
-                                       require_wind_dir=require_wind_dir,
-                                       precip_rate=precip_rate, icing=icing,
-                                       means=means)
 
         return SAM_res
